requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
natsort==8.4.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
perlin-numpy==0.0.1
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
tqdm==4.67.1
